+++ egomimic/rldb/zarr/hdf5_to_zarr.py
@@ -21,7 +21,7 @@ from pathlib import Path
 import numpy as np
 
 from egomimic.rldb.zarr import ZarrWriter
-from egomimic.scripts.eva_process.zarr_utils import EvaHD5Extractor
+from egomimic.scripts.eva_process.eva_utils import EvaHD5Extractor
 from egomimic.rldb.embodiment.eva import Eva
